requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.6.0
python-dotenv==1.0.0

# AWS Services - Compatible with awscli 1.42.72
boto3>=1.35.72
botocore>=1.40.72
s3transfer>=0.14.0,<0.15.0
amazon-braket-sdk>=1.59.0

# LangChain for Agent Framework
langchain>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# Data Processing
numpy>=1.26.3
pandas>=2.1.4

# API
python-multipart==0.0.6
httpx>=0.26.0

mangum==0.17.0